PYTHON/requirements.txt
color_it==2.1.3
colorit==0.1.0
Pillow==9.2.0
requests==2.24.0
